broken up network.go in smaller files
